feat: add consumable indicators and toggles (#1258)
This update improves the way consumables are listed on the actor sheet.  The update adds an indicator (green circle-check icon) to show the consumable that is used when rolling an attack.  Inactive consumables are shown as a grey icon.  You can now switch to a new active consumable be clicking on the appropriate grey icon - no longer need to open the weapon sheet to change the active consumable.  These improvements thanks to the suggestion of @romanK.