feat: Add optional healing dialog to chat context menu (#1864)
This release adds an optional popup dialog when "Apply Healing" chat context is used.  If the (renamed) a setting "Auto Damage and Heal Targets" is set to false, a dialog window will open that allows the owner to adjust how healing is applied to each damage stat.  Thanks to Muji-k for the suggestion.